refactor(synthesis): extract refinement_queue phase to synthesis/refinement_queue.rs
