+++ crates/origin-core/src/refinery/mod.rs
@@ -9,13 +9,12 @@ pub(crate) use helpers::*;
 pub use crate::synthesis::distill::{deep_distill_pages, deep_distill_single, distill_pages};
 
 // Internal re-imports for refinery code that still calls into the moved
-// distillation helpers (apply_merge_by_tier from process_refinement_queue,
-// distill_one_cluster + refine_clusters_with_llm + recompile_single_page from
-// other refinery phases).
-use crate::synthesis::distill::{apply_merge_by_tier, recompile_single_page};
+// distillation helpers (distill_one_cluster + refine_clusters_with_llm +
+// recompile_single_page from other refinery phases).
+use crate::synthesis::distill::recompile_single_page;
+use crate::synthesis::refinement_queue::process_refinement_queue;
 
 use crate::activity::ACTIVITY_GAP_SECS;
-use crate::contradiction::ContradictionResult;
 use crate::db::MemoryDB;
 use crate::error::OriginError;
 use crate::llm_provider::{LlmProvider, LlmRequest};
@@ -1064,115 +1063,6 @@ pub(crate) async fn re_distill_stale_pages(
     Ok(recompiled)
 }
 
-/// Process pending refinement queue items via LLM.
-async fn process_refinement_queue(
-    db: &MemoryDB,
-    llm: Option<&Arc<dyn LlmProvider>>,
-    prompts: &PromptRegistry,
-    tuning: &crate::tuning::RefineryConfig,
-) -> Result<usize, OriginError> {
-    let pending = db.get_pending_refinements().await?;
-    let mut processed = 0usize;
-
-    for proposal in pending.iter().take(tuning.max_proposals_per_steep) {
-        match proposal.action.as_str() {
-            "dedup_merge" => {
-                // Stale v1 proposal — dismiss (distillation handles merges now)
-                db.resolve_refinement(&proposal.id, "dismissed").await?;
-                processed += 1;
-            }
-            "detect_contradiction" => {
-                if let Some(llm) = llm {
-                    let contents = db.get_memory_contents(&proposal.source_ids).await?;
-                    if contents.len() < 2 {
-                        db.resolve_refinement(&proposal.id, "dismissed").await?;
-                        continue;
-                    }
-
-                    let existing_content = contents.get(1).cloned().unwrap_or_default();
-                    let new_content = contents.first().cloned().unwrap_or_default();
-
-                    let response = llm
-                        .generate(LlmRequest {
-                            system_prompt: Some(prompts.detect_contradiction.clone()),
-                            user_prompt: format!(
-                                "Existing: {}\nNew: {}",
-                                existing_content, new_content
-                            ),
-                            max_tokens: 256,
-                            temperature: 0.1,
-                            label: None,
-                            timeout_secs: None,
-                        })
-                        .await;
-
-                    if let Ok(r) = response {
-                        let r = crate::llm_provider::strip_think_tags(&r);
-                        let r = r.trim().to_string();
-                        let result = if r.starts_with("CONTRADICTS:") {
-                            ContradictionResult::Contradicts {
-                                explanation: r
-                                    .strip_prefix("CONTRADICTS:")
-                                    .unwrap_or("")
-                                    .trim()
-                                    .to_string(),
-                            }
-                        } else if r.starts_with("SUPERSEDES:") {
-                            ContradictionResult::Supersedes {
-                                merged_content: r
-                                    .strip_prefix("SUPERSEDES:")
-                                    .unwrap_or("")
-                                    .trim()
-                                    .to_string(),
-                            }
-                        } else {
-                            ContradictionResult::Consistent
-                        };
-
-                        match result {
-                            ContradictionResult::Consistent => {
-                                db.resolve_refinement(&proposal.id, "dismissed").await?;
-                            }
-                            ContradictionResult::Contradicts { explanation } => {
-                                log::info!("[refinery] contradiction detected: {}", explanation);
-                                db.resolve_refinement(&proposal.id, "awaiting_review")
-                                    .await?;
-                            }
-                            ContradictionResult::Supersedes { merged_content } => {
-                                let tier = db.get_highest_tier(&proposal.source_ids).await?;
-                                apply_merge_by_tier(
-                                    db,
-                                    &proposal.source_ids,
-                                    &merged_content,
-                                    &proposal.id,
-                                    &tier,
-                                )
-                                .await?;
-                            }
-                        }
-                        processed += 1;
-                    }
-                }
-            }
-            "suggest_entity" => {
-                // Entity suggestion: payload contains the suggested entity name.
-                // Mark as awaiting_review so the UI can surface it for approval.
-                db.resolve_refinement(&proposal.id, "awaiting_review")
-                    .await?;
-                log::info!(
-                    "[refinery] entity suggestion queued for review: {:?}",
-                    proposal.payload
-                );
-                processed += 1;
-            }
-            _ => {
-                log::debug!("[refinery] unknown action: {}", proposal.action);
-            }
-        }
-    }
-    Ok(processed)
-}
-
 /// Group memories by activity bursts (30-min gap → new burst).
 /// Input memories should be sorted by last_modified (ascending or descending).
 /// Output: groups of references into the input slice, each group is one burst.
@@ -1410,6 +1300,7 @@ mod tests {
     use super::*;
     use crate::db::tests::test_db;
     use crate::sources::{RawDocument, StabilityTier};
+    use crate::synthesis::distill::apply_merge_by_tier;
 
     fn make_memory(source_id: &str, content: &str, memory_type: &str, domain: &str) -> RawDocument {
         RawDocument {
