+++ crates/origin-core/src/synthesis/mod.rs
@@ -9,3 +9,4 @@ pub mod decision_logs;
 pub mod distill;
 pub mod emergence;
 pub mod recaps;
+pub mod refinement_queue;
